Indicate load using LED0
LED0 now gives a basic indication of chip loading. When sv->led_period is 1,
PWM is used to show an analogue value where LED brightness is proportional to
the amount of time the local application cores spend awake.

If any core has an application loaded, the chip's LED is always at least dimly
lit even when all cores are idle. This makes it possible to visually determine
if a machine is being used and see where applications have been loaded.

The familliar sweeping pattern remains to allow distinguishing between idle and
dead cores. This animation, however, now only starts once the machine has
fully booted.

When sv->led_period is 0 or >= 2, the LED behaviour is exactly as it was (i.e.
either undriven or blinking, respectively).