Allocate events before touching shared resources
Because shared resources are seen by other cores, it's best to acquire
events (which are local) before doing anything shared (like allocating a
message from the shared queue or touching a mailbox). Also, doing this
means that the failure mode where a shared message is allocated but an
event is not (in shm_send_msg) can't result in a message getting used
twice at once and can't corrupt the stack of shared messages.

Fixes #108